requirements.txt
python-dotenv>=1.0.0
pyodbc>=4.0.39
google-adk>=1.0.0
litellm>=1.40.0
pydantic>=2.5.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
rich>=13.0.0
prompt_toolkit>=3.0.0
fastapi
uvicorn
httpx
pytest
